backend/pyproject.toml
[project]
name = "permitai-la"
version = "0.1.0"
description = "AI-powered platform to accelerate fire rebuilding permit processes in Los Angeles"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "redis[hiredis]>=5.0",
    "boto3>=1.34",
    "geoalchemy2>=0.14",
    "shapely>=2.0",
    "beautifulsoup4>=4.12",
    "python-jose[cryptography]>=3.3",
    "sentry-sdk[fastapi]>=1.40",
    "structlog>=24.1",
    "firebase-admin>=6.4",
    "twilio>=9.0",
    "python-multipart>=0.0.9",
    "psutil>=5.9",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow>=2.8",
    "apache-airflow-providers-postgres>=5.10",
]
ml = [
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "shap>=0.44",
    "prophet>=1.1",
    "spacy>=3.7",
    "anthropic>=0.18",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "factory-boy>=3.3",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "locust>=2.23",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "T20"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
